requirements.txt
streamlit
spotipy
requests
python-dotenv
agno
googlesearch-python 
pycountry
groq